Fix _is_config attribute for non-config containers
When a YANG definition of a container has the statement "config false;",
the Python bindings generated for this container should have "is_config=False"
in its __init__ method, but it was not the case.

Containers nested in non-config containers correctly inherited the _is_config
attribute from its parent.

Add two test cases to tests/config-false.